app/routers.py
from rest_framework import routers
from app.api_viewsets import UserViewSet, PostViewSet


# Routers provide an easy way of automatically determining the URL conf.
router = routers.DefaultRouter()
router.register(r'users', UserViewSet)
router.register(r'posts', PostViewSet)
