requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
pyhunter
python-dotenv
lorem